A few typos and references to DSG where it was odd
As an addition (and may need to be verified) - in the past extracting images from hackmd was a pain in the backside. I wonder if an additional tip to name images sensibly, put into the image folder in outputs and reference them when they import images into hackmd?

Line 756 RE local files that created in the " /output folder - make a note /shared  folder has its own folder per VM? (You mentioned it before, but might be good to repeat here.

Line 1047 I changed the url to be consistent with the rest of the doc - is that correct?