Add back rule permitting inbound RDP connections to the webapp NSG
The absence of this rule was blocking RDP connections between the APP session host
and the DSVM, since this NSG is applied to the DATA subnet.

There is almost certainly something more fundamentally wrong, but this patch means
that if this branch is deployed (e.g. in a test environment), it won't cause this
issue.