Ensure Firewall is started when updated and when SHM VMs are started (#785)
When a subscription runs out of credit, the Firewall is deallocated. When deallocated no traffic is processed by the Firewall. It also loses its IP configuration(s), so is likely not routable to at all. This PR ensures that the Firewall is started whenever it is configured and whenever the SHM VMs are started.

- Add `Start-Firewall` and `Stop-Firewall` functions to `Deployments.psm1`
- Update `Deploy-Firewall` function to ensure the Firewall is started if it already exists, which also ensures the Firewall exists and is started when updating SHM rules
- Ensure Firewall is started when updating SRE rules
- Ensure Firewall is started when starting SHM VMs with `SHM_Manage_VMs.ps1`
- Add option to deallocate Firewall when stopping SHM VMs with `SHM_Manage_VMs.ps1`
- Add option to pass Exception to be re-thrown by logging function when logging at Fatal level, used by `Deploy-Firewall`

Fixes #783.